# packages/contract/openapi.yaml
openapi: 3.0.3
info:
  title: MeterBase API
  version: "1.0.0"
  description: >
    Usage metering & monetization infrastructure. Ingest usage events, define
    meters, query time-bucketed aggregates, compute cost from pricing models,
    and fire usage-threshold alerts via signed webhooks.

    This document is the source of truth for the v1 REST contract. JSON
    properties are camelCase; the implementation maps to snake_case at the DB
    boundary. All endpoints are scoped to the organization that owns the API key.
servers:
  - url: http://localhost:48888
    description: Local
  - url: https://api.meterbase.example
    description: Production (placeholder)
security:
  - ApiKeyAuth: []
...
paths:
  /v1/events:
    post:
      tags: [Events]
      summary: Ingest one event or a batch of events
      description: >
        Accepts a single event object or an array of events. Events are
        deduplicated on (org, id); a repeated id is accepted (2xx) but not
        double-counted. Aggregation uses the caller-supplied `time` (event
        time), not ingest time. Returns counts of received vs. duplicate events.
      operationId: ingestEvents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Event'
                - type: array
                  items: { $ref: '#/components/schemas/Event' }
                  minItems: 1
            examples:
              single:
                value:
                  id: "evt_01H8XGJ"
                  type: "api_request"
                  source: "api-gateway"
                  subject: "user_123"
                  time: "2026-06-12T10:00:00Z"
                  data: { tokens: 1500, model: "gpt-x" }
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    get:
      tags: [Events]
      summary: List raw events (audit / debug)
      operationId: listEvents
      parameters:
        - { in: query, name: from, schema: { type: string, format: date-time } }
        - { in: query, name: to, schema: { type: string, format: date-time } }
        - { in: query, name: subject, schema: { type: string } }
        - { in: query, name: type, schema: { type: string } }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: A page of raw events
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/StoredEvent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/events/export:
    get:
      tags: [Events]
      summary: Export raw events as a stream (CSV or JSON)
      operationId: exportEvents
      parameters:
        - { in: query, name: format, schema: { type: string, enum: [csv, json], default: json } }
        - { in: query, name: from, schema: { type: string, format: date-time } }
        - { in: query, name: to, schema: { type: string, format: date-time } }
        - { in: query, name: subject, schema: { type: string } }
        - { in: query, name: type, schema: { type: string } }
      responses:
        '200':
          description: Streamed export
          content:
            text/csv: { schema: { type: string } }
            application/json: { schema: { type: array, items: { $ref: '#/components/schemas/StoredEvent' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /v1/customers:
    post:
      tags: [Customers]
      summary: Create a customer
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CustomerCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      tags: [Customers]
      summary: List customers
      operationId: listCustomers
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: A page of customers
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Customer' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/customers/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Customers]
      summary: Get a customer
      operationId: getCustomer
      responses:
        '200':
          description: The customer
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Customers]
      summary: Update a customer
      operationId: updateCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CustomerPatch' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "Pass the API key as a bearer token. Authorization: Bearer <api_key>."
  parameters:
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
      description: Opaque pagination cursor returned as `nextCursor`.
...
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Event:
      type: object
      required: [id, type, subject]
      properties:
        id:
          type: string
          description: Caller-supplied unique id. Dedup key on (org, id).
        type:
          type: string
          description: Must match a meter's eventType to be aggregated.
        source: { type: string }
        subject:
          type: string
          description: Technical consumer (e.g. user_123). Maps to a customer.
        time:
          type: string
          format: date-time
          description: Event time. Defaults to server receipt time if omitted.
        data:
          type: object
          additionalProperties: true
          description: Arbitrary JSON payload; aggregated value & dimensions live here.
    StoredEvent:
      allOf:
        - $ref: '#/components/schemas/Event'
        - type: object
          properties:
            ingestedAt: { type: string, format: date-time }
            customerId: { type: string, nullable: true }
    IngestResponse:
      type: object
      required: [received, duplicates]
      properties:
        received: { type: integer, description: Events accepted and stored. }
        duplicates: { type: integer, description: Events skipped as duplicates. }
...
    CustomerCreate:
      type: object
      required: [externalId]
      properties:
        externalId: { type: string, description: The org's own id for this customer. }
        name: { type: string, nullable: true }
        metadata:
          type: object
          additionalProperties: true
          default: {}
    CustomerPatch:
      type: object
      properties:
        name: { type: string, nullable: true }
        metadata:
          type: object
          additionalProperties: true
    Customer:
      allOf:
        - $ref: '#/components/schemas/CustomerCreate'
        - type: object
          required: [id, createdAt]
          properties:
            id: { type: string }
            createdAt: { type: string, format: date-time }
...
    Page:
      type: object
      properties:
        nextCursor:
          type: string
          nullable: true
          description: Pass as `cursor` to fetch the next page. Null when no more.
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }